Rename RemoveCourse2Activity to RemoveCourseActivity
