+++ app/src/main/res/layout/activity_remove_course.xml
@@ -14,8 +14,8 @@
         <androidx.appcompat.widget.Toolbar
             android:id="@+id/toolbar"
             android:layout_width="match_parent"
-            android:layout_height="?attr/actionBarSize"
-            android:background="?attr/colorPrimary"
+            android:layout_height="1dp"
+            android:background="#ADD8E6"
             app:popupTheme="@style/Theme.GenericTemplate.PopupOverlay" />
 
     </com.google.android.material.appbar.AppBarLayout>
